Split unit test for scandium.
