+++ scandium-core/src/test/java/org/eclipse/californium/scandium/DTLSConnectorTest.java
@@ -42,11 +42,12 @@
  *    Achim Kraus (Bosch Software Innovations GmbH) - remove HELLO_VERIFY_REQUEST
  *                                                    from resumption tests
  *    Achim Kraus (Bosch Software Innovations GmbH) - remove unused sendRecord
+ *    Achim Kraus (Bosch Software Innovations GmbH) - move advanced tests to
+ *                                                    DTLSConnectorAdvancedTest
  ******************************************************************************/
 package org.eclipse.californium.scandium;
 
 import static org.hamcrest.CoreMatchers.*;
-import static org.hamcrest.number.OrderingComparison.greaterThan;
 import static org.junit.Assert.*;
 
 import java.io.IOException;
@@ -86,13 +87,11 @@ import org.eclipse.californium.scandium.dtls.AlertMessage;
 import org.eclipse.californium.scandium.dtls.AlertMessage.AlertDescription;
 import org.eclipse.californium.scandium.dtls.AlertMessage.AlertLevel;
 import org.eclipse.californium.scandium.dtls.CertificateTypeExtension.CertificateType;
-import org.eclipse.californium.scandium.dtls.ClientHandshaker;
 import org.eclipse.californium.scandium.dtls.ClientHello;
 import org.eclipse.californium.scandium.dtls.ClientKeyExchange;
 import org.eclipse.californium.scandium.dtls.CompressionMethod;
 import org.eclipse.californium.scandium.dtls.Connection;
 import org.eclipse.californium.scandium.dtls.ContentType;
-import org.eclipse.californium.scandium.dtls.DTLSFlight;
 import org.eclipse.californium.scandium.dtls.DTLSSession;
 import org.eclipse.californium.scandium.dtls.DtlsTestTools;
 import org.eclipse.californium.scandium.dtls.HandshakeException;
@@ -106,11 +105,6 @@ import org.eclipse.californium.scandium.dtls.InMemorySessionCache;
 import org.eclipse.californium.scandium.dtls.PSKClientKeyExchange;
 import org.eclipse.californium.scandium.dtls.ProtocolVersion;
 import org.eclipse.californium.scandium.dtls.Record;
-import org.eclipse.californium.scandium.dtls.RecordLayer;
-import org.eclipse.californium.scandium.dtls.ResumingClientHandshaker;
-import org.eclipse.californium.scandium.dtls.ResumingServerHandshaker;
-import org.eclipse.californium.scandium.dtls.ServerHandshaker;
-import org.eclipse.californium.scandium.dtls.SessionAdapter;
 import org.eclipse.californium.scandium.dtls.SessionId;
 import org.eclipse.californium.scandium.dtls.cipher.CipherSuite;
 import org.eclipse.californium.scandium.dtls.pskstore.InMemoryPskStore;
@@ -240,7 +234,7 @@ public class DTLSConnectorTest {
 		clientRawDataChannel = new LatchDecrementingRawDataChannel();
 	}
 
-	public void autoResumeSetUp(int timeout) throws Exception {
+	private void autoResumeSetUp(int timeout) throws Exception {
 		cleanUp();
 		serverSessionCache.establishedSessionCounter.set(0);
 		pskStoreLatency = 0;
@@ -679,645 +673,6 @@ public class DTLSConnectorTest {
 			assertThat("Expected ALERT message from server",
 					record.getType(), is(ContentType.ALERT));
 
-		} finally {
-			rawClient.stop();
-		}
-	}
-
-	@Test
-	public void testServerReceivingMessagesInBadOrderDuringHandshake() throws Exception {
-		// Configure and create UDP connector
-		RecordCollectorDataHandler collector = new RecordCollectorDataHandler();
-		UdpConnector rawClient = new UdpConnector(clientEndpoint, collector);
-		try {
-
-			// Start connector
-			rawClient.start();
-			clientEndpoint = new InetSocketAddress(rawClient.socket.getLocalAddress(), rawClient.socket.getLocalPort());
-			LatchSessionListener sessionListener = new LatchSessionListener();
-
-			// Create handshaker with ReverseRecordLayer to send message in bad
-			// order.
-			ClientHandshaker clientHandshaker = new ClientHandshaker(new DTLSSession(serverEndpoint, true),
-					new ReverseRecordLayer(rawClient), sessionListener, clientConfig, 1280);
-
-			// Start handshake (Send CLIENT HELLO)
-			clientHandshaker.startHandshake();
-
-			// Wait to receive response (should be HELLO VERIFY REQUEST)
-			List<Record> rs = collector.waitForFlight(MAX_TIME_TO_WAIT_SECS, TimeUnit.SECONDS);
-			assertNotNull("timeout", rs); // check there is no timeout
-			// Handle and answer (CLIENT HELLO with cookie)
-			for (Record r : rs) {
-				clientHandshaker.processMessage(r);
-			}
-
-			// Wait for response (SERVER_HELLO, CERTIFICATE, ... , SERVER_DONE)
-			rs = collector.waitForFlight(MAX_TIME_TO_WAIT_SECS, TimeUnit.SECONDS);
-			assertNotNull("timeout", rs);
-			// Handle and answer (FINISHED, CHANGE CIPHER SPEC, ...,CERTIFICATE)
-			for (Record r : rs) {
-				clientHandshaker.processMessage(r);
-			}
-
-			// Wait to receive response (should be CHANGE CIPHER SPEC, FINISHED)
-			rs = collector.waitForFlight(MAX_TIME_TO_WAIT_SECS, TimeUnit.SECONDS);
-			assertNotNull("timeout", rs);
-			// Handle it
-			for (Record r : rs) {
-				clientHandshaker.processMessage(r);
-			}
-
-			// Ensure handshake is successfully done
-			assertTrue("handshake failed",
-					sessionListener.waitForSessionEstablished(MAX_TIME_TO_WAIT_SECS, TimeUnit.SECONDS));
-		} finally {
-			rawClient.stop();
-		}
-	}
-
-	@Test
-	public void testClientReceivingMessagesInBadOrderDuringHandshake() throws Exception {
-		// Configure UDP connector we will use as Server
-		RecordCollectorDataHandler collector = new RecordCollectorDataHandler();
-		UdpConnector rawServer = new UdpConnector(new InetSocketAddress(0), collector);
-
-		try {
-			// Start connector (Server)
-			rawServer.start();
-			InetSocketAddress rawServerEndpoint = new InetSocketAddress("localhost", rawServer.socket.getLocalPort());
-			LatchSessionListener sessionListener = new LatchSessionListener();
-
-			// Start the client
-			CountDownLatch latch = new CountDownLatch(1);
-			clientRawDataChannel.setLatch(latch);
-			client.setRawDataReceiver(clientRawDataChannel);
-			client.start();
-			clientEndpoint = client.getAddress();
-			RawData data = RawData.outbound("Hello World".getBytes(), new AddressEndpointContext(rawServerEndpoint), null, false);
-			client.send(data);
-
-			// Create server handshaker
-			ServerHandshaker serverHandshaker = new ServerHandshaker(new DTLSSession(clientEndpoint, false, 1),
-					new ReverseRecordLayer(rawServer), sessionListener, serverConfig, 1280);
-
-			// Wait to receive response (should be CLIENT HELLO, flight 3)
-			List<Record> rs = collector.waitForFlight(MAX_TIME_TO_WAIT_SECS, TimeUnit.SECONDS);
-			assertNotNull("timeout", rs); // check there is no timeout
-			// Handle and answer (should be CERTIFICATE, ... SERVER HELLO DONE, flight 4)
-			for (Record r : rs) {
-				serverHandshaker.processMessage(r);
-			}
-
-			// Wait for receive response (CERTIFICATE, ... , FINISHED, flight 5)
-			rs = collector.waitForFlight(MAX_TIME_TO_WAIT_SECS, TimeUnit.SECONDS);
-			assertNotNull("timeout", rs);
-			// Handle and answer (should be CCS, FINISHED, flight 6)
-			for (Record r : rs) {
-				serverHandshaker.processMessage(r);
-			}
-
-			// Ensure handshake is successfully done
-			assertTrue("handshake failed",
-					sessionListener.waitForSessionEstablished(MAX_TIME_TO_WAIT_SECS, TimeUnit.SECONDS));
-		} finally {
-			rawServer.stop();
-		}
-	}
-
-	@Test
-	public void testServerResumeReceivingMessagesInBadOrderDuringHandshake() throws Exception {
-		// Configure and create UDP connector
-		RecordCollectorDataHandler collector = new RecordCollectorDataHandler();
-		UdpConnector rawClient = new UdpConnector(clientEndpoint, collector);
-		ReverseRecordLayer clientRecordLayer = new ReverseRecordLayer(rawClient);
-		DTLSSession clientSession = new DTLSSession(serverEndpoint, true);
-		try {
-
-			// Start connector
-			rawClient.start();
-			clientEndpoint = new InetSocketAddress(rawClient.socket.getLocalAddress(), rawClient.socket.getLocalPort());
-			LatchSessionListener sessionListener = new LatchSessionListener();
-
-			// Create handshaker
-			ClientHandshaker clientHandshaker = new ClientHandshaker(clientSession, clientRecordLayer, sessionListener,
-					clientConfig, 1280);
-
-			// Start 1. handshake (Send CLIENT HELLO)
-			clientHandshaker.startHandshake();
-
-			// Wait to receive response (should be HELLO VERIFY REQUEST)
-			List<Record> rs = collector.waitForFlight(MAX_TIME_TO_WAIT_SECS, TimeUnit.SECONDS);
-			assertNotNull("timeout", rs); // check there is no timeout
-			// Handle and answer (CLIENT HELLO with cookie)
-			for (Record r : rs) {
-				clientHandshaker.processMessage(r);
-			}
-
-			// Wait for response (SERVER_HELLO, CERTIFICATE, ... , SERVER_DONE)
-			rs = collector.waitForFlight(MAX_TIME_TO_WAIT_SECS, TimeUnit.SECONDS);
-			assertNotNull("timeout", rs);
-			// Handle and answer (CERTIFICATE, CHANGE CIPHER SPEC, ...,
-			// FINISHED)
-			for (Record r : rs) {
-				clientHandshaker.processMessage(r);
-			}
-
-			// Wait to receive response from server (should be CHANGE CIPHER
-			// SPEC, FINISHED)
-			rs = collector.waitForFlight(MAX_TIME_TO_WAIT_SECS, TimeUnit.SECONDS);
-			assertNotNull("timeout", rs);
-			// Handle (CHANGE CIPHER SPEC, FINISHED)
-			for (Record r : rs) {
-				clientHandshaker.processMessage(r);
-			}
-
-			// Ensure handshake is successfully done
-			assertTrue("handshake failed",
-					sessionListener.waitForSessionEstablished(MAX_TIME_TO_WAIT_SECS, TimeUnit.SECONDS));
-
-			// Create resume handshaker
-			sessionListener = new LatchSessionListener();
-			DTLSSession resumableSession = new DTLSSession(clientSession.getSessionIdentifier(), serverEndpoint,
-					clientSession.getSessionTicket(), 0);
-			ResumingClientHandshaker resumingClientHandshaker = new ResumingClientHandshaker(resumableSession,
-					clientRecordLayer, sessionListener, clientConfig, 1280);
-
-
-			// Start resuming handshake (Send CLIENT HELLO, additional flight)
-			resumingClientHandshaker.startHandshake();
-
-			// Wait to receive response (should be SERVER_HELLO, CHANGE CIPHER SPEC, FINISHED, flight 2)
-			rs = collector.waitForFlight(MAX_TIME_TO_WAIT_SECS, TimeUnit.SECONDS);
-			assertNotNull("timeout", rs); // check there is no timeout
-			assertEquals(3, rs.size());
-
-			// create server session listener to ensure, that server finish also the handshake
-			Connection con = serverConnectionStore.get(clientEndpoint);
-			assertNotNull(con);
-			Handshaker handshaker = con.getOngoingHandshake();
-			assertNotNull(handshaker);
-			LatchSessionListener serverSessionListener = new LatchSessionListener();
-			handshaker.addSessionListener(serverSessionListener);
-
-			// Handle and answer ( CHANGE CIPHER SPEC, FINISHED, flight 3)
-			for (Record r : rs) {
-				resumingClientHandshaker.processMessage(r);
-			}
-
-			// Ensure handshake is successfully done
-			assertTrue("client handshake failed",
-					sessionListener.waitForSessionEstablished(MAX_TIME_TO_WAIT_SECS, TimeUnit.SECONDS));
-
-			assertTrue("server handshake failed",
-					serverSessionListener.waitForSessionEstablished(MAX_TIME_TO_WAIT_SECS, TimeUnit.SECONDS));
-		} finally {
-			rawClient.stop();
-		}
-	}
-
-	@Test
-	public void testClientResumeReceivingMessagesInBadOrderDuringHandshake() throws Exception {
-		// Configure UDP connector we will use as Server
-		RecordCollectorDataHandler collector = new RecordCollectorDataHandler();
-		UdpConnector rawServer = new UdpConnector(new InetSocketAddress(0), collector);
-		ReverseRecordLayer serverRecordLayer = new ReverseRecordLayer(rawServer);
-
-		try {
-			// Start connector (Server)
-			rawServer.start();
-
-			InetSocketAddress rawServerEndpoint = new InetSocketAddress("localhost", rawServer.socket.getLocalPort());
-			LatchSessionListener sessionListener = new LatchSessionListener();
-
-			// Start the client
-			client.setRawDataReceiver(clientRawDataChannel);
-			client.start();
-			clientEndpoint = client.getAddress();
-			RawData data = RawData.outbound("Hello World".getBytes(), new AddressEndpointContext(rawServerEndpoint), null, false);
-			client.send(data);
-
-			// Create server handshaker
-			DTLSSession serverSession = new DTLSSession(clientEndpoint, false, 1);
-			ServerHandshaker serverHandshaker = new ServerHandshaker(serverSession, serverRecordLayer, sessionListener,
-					serverConfig, 1280);
-
-			// 1. handshake
-			// Wait to receive response (should be CLIENT HELLO)
-			List<Record> rs = collector.waitForFlight(MAX_TIME_TO_WAIT_SECS, TimeUnit.SECONDS);
-			assertNotNull("timeout", rs); // check there is no timeout
-			// Handle and answer (should be CERTIFICATE, ... SERVER HELLO DONE)
-			for (Record r : rs) {
-				serverHandshaker.processMessage(r);
-			}
-
-			// Wait to receive response (CERTIFICATE, ... , FINISHED)
-			rs = collector.waitForFlight(MAX_TIME_TO_WAIT_SECS, TimeUnit.SECONDS);
-			assertNotNull("timeout", rs);
-			for (Record r : rs) {
-				serverHandshaker.processMessage(r);
-			}
-
-			// Ensure handshake is successfully done
-			assertTrue("handshake failed",
-					sessionListener.waitForSessionEstablished(MAX_TIME_TO_WAIT_SECS, TimeUnit.SECONDS));
-
-			// application data
-			rs = collector.waitForFlight(MAX_TIME_TO_WAIT_SECS, TimeUnit.SECONDS);
-			assertNotNull("timeout", rs);
-			assertEquals(1, rs.size());
-
-			sessionListener = new LatchSessionListener();
-			DTLSSession resumableSession = new DTLSSession(serverSession.getSessionIdentifier(), clientEndpoint,
-					serverSession.getSessionTicket(), 0);
-			ResumingServerHandshaker resumingServerHandshaker = new ResumingServerHandshaker(0, resumableSession,
-					serverRecordLayer, sessionListener, serverConfig, 1280);
-
-			// force resuming handshake
-			client.forceResumeSessionFor(rawServerEndpoint);
-			data = RawData.outbound("Hello Again".getBytes(), new AddressEndpointContext(rawServerEndpoint), null, false);
-			client.send(data);
-
-			// Wait to receive response (should be CLIENT HELLO, flight 1)
-			rs = collector.waitForFlight(MAX_TIME_TO_WAIT_SECS, TimeUnit.SECONDS);
-			assertNotNull("timeout", rs);
-			assertEquals(1, rs.size());
-			// Handle and answer (should be SERVER HELLO, CCS, FINISHED, flight 2).
-			// Note: additional flight used by scandium
-			for (Record r : rs) {
-				resumingServerHandshaker.processMessage(r);
-			}
-
-			// Wait to receive response (CCS, client FINISHED, flight 3) + (application data)
-			List<Record> drops = collector.waitForFlight(MAX_TIME_TO_WAIT_SECS, TimeUnit.SECONDS);
-			assertNotNull("timeout", drops);
-			assertEquals(3, drops.size());
-			// remove application data, not retransmitted!
-			drops.remove(2); 
-			
-			// drop last flight 3, server resends flight 2
-			serverRecordLayer.resendLastFlight();
-
-			// Wait to receive response (CCS, client FINISHED, flight 3)
-			// ("application data" doesn't belong to flight)
-			rs = collector.waitForFlight(MAX_TIME_TO_WAIT_SECS, TimeUnit.SECONDS);
-			assertNotNull("timeout", rs);
-			assertEquals(2, rs.size());
-			assertFlightRecordsRetransmitted(drops, rs);
-			for (Record r : rs) {
-				resumingServerHandshaker.processMessage(r);
-			}
-			assertTrue("handshake failed",
-					sessionListener.waitForSessionEstablished(MAX_TIME_TO_WAIT_SECS, TimeUnit.SECONDS));
-
-		} finally {
-			rawServer.stop();
-		}
-	}
-	
-	
-	/**
-	 * Test retransmission of last flight.
-	 * 
-	 * RFC6347, section 4.2.4, fig. 2
-	 * 
-	 * "testFinishedMessageRetransmission" drops the first transmission of
-	 * flight 5 to test, if flight 5 is retransmitted. But flight 5 is just
-	 * usual retransmission, the special case is flight 6. Therefore this test
-	 * drops the 1. transmission of flight 6 to check, if retransmission of
-	 * flight 5 triggers the retransmission of flight 6.
-	 */
-	@Test
-	public void testServerFinishedMessageRetransmission() throws Exception {
-		// Configure and create UDP connector
-		RecordCollectorDataHandler collector = new RecordCollectorDataHandler();
-		UdpConnector rawClient = new UdpConnector(clientEndpoint, collector);
-		SimpleRecordLayer clientRecordLayer = new SimpleRecordLayer(rawClient);
-		try {
-
-			// Start connector
-			rawClient.start();
-			clientEndpoint = new InetSocketAddress(rawClient.socket.getLocalAddress(), rawClient.socket.getLocalPort());
-			LatchSessionListener sessionListener = new LatchSessionListener();
-
-			// Create handshaker
-			ClientHandshaker clientHandshaker = new ClientHandshaker(new DTLSSession(serverEndpoint, true),
-					clientRecordLayer, sessionListener, clientConfig, 1280);
-
-			// Start handshake (Send CLIENT HELLO, flight 1)
-			clientHandshaker.startHandshake();
-
-			// Wait to receive response (should be HELLO VERIFY REQUEST, flight 2)
-			List<Record> rs = collector.waitForFlight(MAX_TIME_TO_WAIT_SECS, TimeUnit.SECONDS);
-			assertNotNull("timeout", rs); // check there is no timeout
-			// Handle and answer (CLIENT HELLO with cookie, flight 3)
-			for (Record r : rs) {
-				clientHandshaker.processMessage(r);
-			}
-
-			// Wait for response (SERVER_HELLO, CERTIFICATE, ... , SERVER_DONE, flight 4)
-			rs = collector.waitForFlight(MAX_TIME_TO_WAIT_SECS, TimeUnit.SECONDS);
-			assertNotNull("timeout", rs);
-			// Handle and answer (CERTIFICATE, CHANGE CIPHER SPEC, ..., FINISHED, flight 5)
-			for (Record r : rs) {
-				clientHandshaker.processMessage(r);
-			}
-
-			// Wait to receive response from server (should be CHANGE CIPHER SPEC, FINISHED, flight 6)
-			List<Record> drops = collector.waitForFlight(MAX_TIME_TO_WAIT_SECS, TimeUnit.SECONDS);
-			assertNotNull("timeout", drops);
-			// Ignore the receive response, client resends flight 5
-			clientRecordLayer.resendLastFlight();
-
-			// Wait for retransmission (should be CHANGE CIPHER SPEC, FINISHED, flight 6)
-			rs = collector.waitForFlight(MAX_TIME_TO_WAIT_SECS, TimeUnit.SECONDS);
-			assertNotNull("timeout", rs);
-			assertFlightRecordsRetransmitted(drops, rs);
-			for (Record r : rs) {
-				clientHandshaker.processMessage(r);
-			}
-
-			// Ensure handshake is successfully done
-			assertTrue("handshake failed",
-					sessionListener.waitForSessionEstablished(MAX_TIME_TO_WAIT_SECS, TimeUnit.SECONDS));
-		} finally {
-			rawClient.stop();
-		}
-	}
-
-	/**
-	 * Test retransmission of last flight of resuming handshake.
-	 * 
-	 * RFC6347, section 4.2.4, fig. 2
-	 * 
-	 * "testResumeFinishedMessageRetransmission" drops the first transmission of
-	 * flight 2 to test, if flight 2 is retransmitted. But flight 2 is just
-	 * usual retransmission, the special case is flight 3. Therefore this test
-	 * drops the 1. transmission of flight 3 to check, if retransmission of
-	 * flight 2 triggers the retransmission of flight 3.
-	 */
-	@Test
-	public void testResumeClientFinishedMessageRetransmission() throws Exception {
-		// Configure UDP connector we will use as Server
-		RecordCollectorDataHandler collector = new RecordCollectorDataHandler();
-		UdpConnector rawServer = new UdpConnector(new InetSocketAddress(0), collector);
-		SimpleRecordLayer serverRecordLayer = new SimpleRecordLayer(rawServer);
-
-		try {
-			// Start connector (Server)
-			rawServer.start();
-
-			InetSocketAddress rawServerEndpoint = new InetSocketAddress("localhost", rawServer.socket.getLocalPort());
-			LatchSessionListener sessionListener = new LatchSessionListener();
-
-			// Start the client
-			client.setRawDataReceiver(clientRawDataChannel);
-			client.start();
-			clientEndpoint = client.getAddress();
-			RawData data = RawData.outbound("Hello World".getBytes(), new AddressEndpointContext(rawServerEndpoint), null, false);
-			client.send(data);
-
-			// Create server handshaker
-			DTLSSession serverSession = new DTLSSession(clientEndpoint, false, 1);
-			ServerHandshaker serverHandshaker = new ServerHandshaker(serverSession, serverRecordLayer, sessionListener,
-					serverConfig, 1280);
-
-			// 1. handshake
-			// Wait to receive response (should be CLIENT HELLO)
-			List<Record> rs = collector.waitForFlight(MAX_TIME_TO_WAIT_SECS, TimeUnit.SECONDS);
-			assertNotNull("timeout", rs); // check there is no timeout
-			// Handle and answer (should be CERTIFICATE, ... SERVER HELLO DONE)
-			for (Record r : rs) {
-				serverHandshaker.processMessage(r);
-			}
-
-			// Wait to receive response (CERTIFICATE, ... , FINISHED)
-			rs = collector.waitForFlight(MAX_TIME_TO_WAIT_SECS, TimeUnit.SECONDS);
-			assertNotNull("timeout", rs);
-			for (Record r : rs) {
-				serverHandshaker.processMessage(r);
-			}
-
-			// Ensure handshake is successfully done
-			assertTrue("handshake failed",
-					sessionListener.waitForSessionEstablished(MAX_TIME_TO_WAIT_SECS, TimeUnit.SECONDS));
-
-			// application data
-			rs = collector.waitForFlight(MAX_TIME_TO_WAIT_SECS, TimeUnit.SECONDS);
-			assertNotNull("timeout", rs);
-			assertEquals(1, rs.size());
-
-			sessionListener = new LatchSessionListener();
-			DTLSSession resumableSession = new DTLSSession(serverSession.getSessionIdentifier(), clientEndpoint,
-					serverSession.getSessionTicket(), 0);
-			ResumingServerHandshaker resumingServerHandshaker = new ResumingServerHandshaker(0, resumableSession,
-					serverRecordLayer, sessionListener, serverConfig, 1280);
-
-			// force resuming handshake
-			client.forceResumeSessionFor(rawServerEndpoint);
-			data = RawData.outbound("Hello Again".getBytes(), new AddressEndpointContext(rawServerEndpoint), null, false);
-			client.send(data);
-
-			// Wait to receive response (should be CLIENT HELLO, flight 1)
-			rs = collector.waitForFlight(MAX_TIME_TO_WAIT_SECS, TimeUnit.SECONDS);
-			assertNotNull("timeout", rs);
-			assertEquals(1, rs.size());
-			// Handle and answer (should be SERVER HELLO, CCS, FINISHED, VERIFY REQUEST, flight 2).
-			// Note: additional flight used by scandium
-			for (Record r : rs) {
-				resumingServerHandshaker.processMessage(r);
-			}
-
-			// Wait to receive response (CCS, client FINISHED, flight 3) + (application data)
-			List<Record> drops = collector.waitForFlight(MAX_TIME_TO_WAIT_SECS, TimeUnit.SECONDS);
-			assertNotNull("timeout", drops);
-			assertEquals(3, drops.size());
-			// remove application data, not retransmitted!
-			drops.remove(2); 
-			
-			// drop last flight 3, server resends flight 2
-			serverRecordLayer.resendLastFlight();
-
-			// Wait to receive response (CCS, client FINISHED, flight 3)
-			// ("application data" doesn't belong to flight)
-			rs = collector.waitForFlight(MAX_TIME_TO_WAIT_SECS, TimeUnit.SECONDS);
-			assertNotNull("timeout", rs);
-			assertEquals(2, rs.size());
-			assertFlightRecordsRetransmitted(drops, rs);
-			for (Record r : rs) {
-				resumingServerHandshaker.processMessage(r);
-			}
-			assertTrue("handshake failed",
-					sessionListener.waitForSessionEstablished(MAX_TIME_TO_WAIT_SECS, TimeUnit.SECONDS));
-
-		} finally {
-			rawServer.stop();
-		}
-	}
-
-	/**
-	 * Test retransmission of flight before last flight.
-	 * 
-	 * RFC6347, section 4.2.4, fig. 2
-	 * 
-	 * Drops the first transmission of flight 5 to test, if flight 5 is retransmitted.
-	 * Usual retransmission, the special case is flight 6, see "testServerFinishedMessageRetransmission". 
-	 */
-	@Test
-	public void testFinishedMessageRetransmission() throws Exception {
-		// Configure UDP connector we will use as Server
-		RecordCollectorDataHandler collector = new RecordCollectorDataHandler();
-		UdpConnector rawServer = new UdpConnector(new InetSocketAddress(0), collector);
-
-		try {
-			// Start connector (Server)
-			rawServer.start();
-			InetSocketAddress rawServerEndpoint = new InetSocketAddress("localhost", rawServer.socket.getLocalPort());
-			LatchSessionListener sessionListener = new LatchSessionListener();
-
-			// Start the client
-			CountDownLatch latch = new CountDownLatch(1);
-			clientRawDataChannel.setLatch(latch);
-			client.setRawDataReceiver(clientRawDataChannel);
-			client.start();
-			clientEndpoint = client.getAddress();
-			RawData data = RawData.outbound("Hello World".getBytes(), new AddressEndpointContext(rawServerEndpoint), null, false);
-			client.send(data);
-
-			// Create server handshaker
-			ServerHandshaker serverHandshaker = new ServerHandshaker(new DTLSSession(clientEndpoint, false, 1),
-					new SimpleRecordLayer(rawServer), sessionListener, serverConfig, 1280);
-
-			// Wait to receive response (should be CLIENT HELLO, flight 3)
-			List<Record> rs = collector.waitForFlight(MAX_TIME_TO_WAIT_SECS, TimeUnit.SECONDS);
-			assertNotNull("timeout", rs); // check there is no timeout
-			// Handle and answer (should be CERTIFICATE, ... SERVER HELLO DONE, flight 4)
-			for (Record r : rs) {
-				serverHandshaker.processMessage(r);
-			}
-
-			// Ignore transmission (CERTIFICATE, ... , FINISHED, flight 5)
-			List<Record> drops = collector.waitForFlight(MAX_TIME_TO_WAIT_SECS, TimeUnit.SECONDS);
-			assertNotNull("timeout", drops);
-
-			// Wait for retransmission (CERTIFICATE, ... , FINISHED, flight 5)
-			rs = collector.waitForFlight(MAX_TIME_TO_WAIT_SECS, TimeUnit.SECONDS);
-			assertNotNull("timeout", rs);
-			assertFlightRecordsRetransmitted(drops, rs);
-			// Handle and answer (should be CCS, FINISHED, flight 6)
-			for (Record r : rs) {
-				serverHandshaker.processMessage(r);
-			}
-
-			// Ensure handshake is successfully done
-			assertTrue("handshake failed",
-					sessionListener.waitForSessionEstablished(MAX_TIME_TO_WAIT_SECS, TimeUnit.SECONDS));
-		} finally {
-			rawServer.stop();
-		}
-	}
-
-	/**
-	 * Test retransmission of flight before last flight of resuming handshake.
-	 * 
-	 * RFC6347, section 4.2.4, fig. 2
-	 * 
-	 * Drops the first transmission of flight 2 to test, if flight 2 is
-	 * retransmitted. Flight 2 is just usual retransmission, the special case is
-	 * flight 3, see "testResumeClientFinishedMessageRetransmission". Note:
-	 * scandium uses a additional HELLO VERIFY REQUEST. May be optimized in the
-	 * future.
-	 */
-	@Test
-	public void testResumeFinishedMessageRetransmission() throws Exception {
-		// Configure and create UDP connector
-		RecordCollectorDataHandler collector = new RecordCollectorDataHandler();
-		UdpConnector rawClient = new UdpConnector(clientEndpoint, collector);
-		SimpleRecordLayer clientRecordLayer = new SimpleRecordLayer(rawClient);
-		DTLSSession clientSession = new DTLSSession(serverEndpoint, true);
-		try {
-
-			// Start connector
-			rawClient.start();
-			clientEndpoint = new InetSocketAddress(rawClient.socket.getLocalAddress(), rawClient.socket.getLocalPort());
-			LatchSessionListener sessionListener = new LatchSessionListener();
-
-			// Create handshaker
-			ClientHandshaker clientHandshaker = new ClientHandshaker(clientSession, clientRecordLayer, sessionListener,
-					clientConfig, 1280);
-
-			// Start 1. handshake (Send CLIENT HELLO)
-			clientHandshaker.startHandshake();
-
-			// Wait to receive response (should be HELLO VERIFY REQUEST)
-			List<Record> rs = collector.waitForFlight(MAX_TIME_TO_WAIT_SECS, TimeUnit.SECONDS);
-			assertNotNull("timeout", rs); // check there is no timeout
-			// Handle and answer (CLIENT HELLO with cookie)
-			for (Record r : rs) {
-				clientHandshaker.processMessage(r);
-			}
-
-			// Wait for response (SERVER_HELLO, CERTIFICATE, ... , SERVER_DONE)
-			rs = collector.waitForFlight(MAX_TIME_TO_WAIT_SECS, TimeUnit.SECONDS);
-			assertNotNull("timeout", rs);
-			// Handle and answer (CERTIFICATE, CHANGE CIPHER SPEC, ...,
-			// FINISHED)
-			for (Record r : rs) {
-				clientHandshaker.processMessage(r);
-			}
-
-			// Wait to receive response from server (should be CHANGE CIPHER
-			// SPEC, FINISHED)
-			rs = collector.waitForFlight(MAX_TIME_TO_WAIT_SECS, TimeUnit.SECONDS);
-			assertNotNull("timeout", rs);
-			// Handle (CHANGE CIPHER SPEC, FINISHED)
-			for (Record r : rs) {
-				clientHandshaker.processMessage(r);
-			}
-
-			// Ensure handshake is successfully done
-			assertTrue("handshake failed",
-					sessionListener.waitForSessionEstablished(MAX_TIME_TO_WAIT_SECS, TimeUnit.SECONDS));
-
-			// Create resume handshaker
-			sessionListener = new LatchSessionListener();
-			DTLSSession resumableSession = new DTLSSession(clientSession.getSessionIdentifier(), serverEndpoint,
-					clientSession.getSessionTicket(), 0);
-			ResumingClientHandshaker resumingClientHandshaker = new ResumingClientHandshaker(resumableSession,
-					clientRecordLayer, sessionListener, clientConfig, 1280);
-
-			// Start resuming handshake (Send CLIENT HELLO, additional flight)
-			resumingClientHandshaker.startHandshake();
-
-			// Wait to receive response (should be SERVER_HELLO, CHANGE CIPHER SPEC, FINISHED, fight 2)
-			List<Record> drops = collector.waitForFlight(MAX_TIME_TO_WAIT_SECS, TimeUnit.SECONDS);
-			assertNotNull("timeout", drops); // check there is no timeout
-			assertEquals(3, drops.size());
-
-			// create server session listener to ensure, that server finish also the handshake
-			Connection con = serverConnectionStore.get(clientEndpoint);
-			assertNotNull(con);
-			Handshaker handshaker = con.getOngoingHandshake();
-			assertNotNull(handshaker);
-			LatchSessionListener serverSessionListener = new LatchSessionListener();
-			handshaker.addSessionListener(serverSessionListener);
-			
-			// drop it, force retransmission (should be SERVER_HELLO, CHANGE CIPHER SPEC, FINISHED, fight 2)
-			rs = collector.waitForFlight(MAX_TIME_TO_WAIT_SECS, TimeUnit.SECONDS);
-			assertNotNull("timeout", rs); // check there is no timeout
-			assertEquals(3, rs.size());
-			assertFlightRecordsRetransmitted(drops, rs);
-			// Handle and answer ( CHANGE CIPHER SPEC, FINISHED, flight 3)
-			for (Record r : rs) {
-				resumingClientHandshaker.processMessage(r);
-			}
-
-			// Ensure handshake is successfully done
-			assertTrue("client handshake failed",
-					sessionListener.waitForSessionEstablished(MAX_TIME_TO_WAIT_SECS, TimeUnit.SECONDS));
-
-			assertTrue("server handshake failed",
-					serverSessionListener.waitForSessionEstablished(MAX_TIME_TO_WAIT_SECS, TimeUnit.SECONDS));
 		} finally {
 			rawClient.stop();
 		}
@@ -1893,19 +1248,6 @@ public class DTLSConnectorTest {
 		}
 	}
 
-	private void assertFlightRecordsRetransmitted(final List<Record> flight1, final List<Record> flight2) {
-
-		// assert that flights contains the same number of records.
-		assertThat("retransmitted flight has different number of records", flight2.size(), is(flight1.size()));
-		for (int index=0; index < flight1.size(); ++index) {
-			Record record1 = flight1.get(index);
-			Record record2 = flight2.get(index);
-			assertThat("retransmitted flight record has different epoch", record2.getEpoch(), is(record1.getEpoch()));
-			assertThat("retransmitted flight record has different type", record2.getType(), is(record1.getType()));
-			assertThat("retransmitted flight record has no newer seqn", record2.getSequenceNumber(), is(greaterThan(record1.getSequenceNumber())));
-		}
-	}
-
 	@Test
 	public void testGetMaximumTransmissionUnitReturnsDefaultValue() {
 		// given a connector that has not been started yet
@@ -2106,33 +1448,6 @@ public class DTLSConnectorTest {
 		public List<Record> waitForRecords(long timeout, TimeUnit unit) throws InterruptedException {
 			return records.poll(timeout, unit);
 		}
-
-		public List<Record> waitForFlight(long timeout, TimeUnit unit) throws InterruptedException {
-			List<Record> received = waitForRecords(timeout, unit);
-			if (null != received) {
-				received = new ArrayList<Record>(received);
-				List<Record> next;
-				if (null != (next = waitForRecords(200, TimeUnit.MILLISECONDS))) {
-					received.addAll(next);
-				}
-			}
-			return received;
-		}
-	};
-
-	private static class LatchSessionListener extends SessionAdapter {
-
-		private CountDownLatch latch = new CountDownLatch(1);
-
-		@Override
-		public void sessionEstablished(Handshaker handshaker, DTLSSession establishedSession)
-				throws HandshakeException {
-			latch.countDown();
-		}
-
-		public boolean waitForSessionEstablished(long timeout, TimeUnit unit) throws InterruptedException {
-			return latch.await(timeout, unit);
-		}
 	};
 
 	private void givenAnIncompleteHandshake() throws Exception {
@@ -2217,64 +1532,6 @@ public class DTLSConnectorTest {
 			}
 		}
 	}
-
-	public static class SimpleRecordLayer implements RecordLayer {
-
-		protected final UdpConnector connector;
-		protected volatile DTLSFlight lastFlight;
-
-		public SimpleRecordLayer(UdpConnector connector) {
-			this.connector = connector;
-		}
-
-		@Override
-		public void sendFlight(DTLSFlight flight) {
-			lastFlight = flight;
-			for (Record record : flight.getMessages()) {
-				try {
-					connector.sendRecord(record.getPeerAddress(), record.toByteArray());
-				} catch (IOException e) {
-					throw new RuntimeException(e);
-				}
-			}
-		}
-
-		public void resendLastFlight() throws GeneralSecurityException {
-			DTLSFlight flight = lastFlight;
-			if (null == flight) {
-				throw new IllegalStateException("no last flight!");
-			}
-			flight.incrementTries();
-			flight.setNewSequenceNumbers();
-			sendFlight(flight);
-		}
-
-		@Override
-		public void cancelRetransmissions() {
-		}
-	};
-
-	public static class ReverseRecordLayer extends SimpleRecordLayer {
-
-		public ReverseRecordLayer(UdpConnector connector) {
-			super(connector);
-		}
-
-		@Override
-		public void sendFlight(DTLSFlight flight) {
-			lastFlight = flight;
-			List<Record> messages = flight.getMessages();
-			for (int index = messages.size() - 1; index >= 0; index--) {
-				try {
-					Record record = messages.get(index);
-					LOGGER.info("Reverse send {}: {}", index, record);
-					connector.sendRecord(record.getPeerAddress(), record.toByteArray());
-				} catch (IOException e) {
-					throw new RuntimeException(e);
-				}
-			}
-		}
-	};
 
 	private static class SimpleRawDataChannel implements RawDataChannel {
